Update sidebar and wallet pages
- Update Wallets components folder structure
- Add active sidebar links
- Update UI
- Bug fix
