+++ app/(dashboard)/layout.js
@@ -8,10 +8,10 @@ export const metadata = {
 
 export default function DashboardLayout({ children }) {
   return (
-    <div className="relative flex gap-5">
+    <div className="hero-bg relative flex gap-5">
       <Sidebar />
 
-      <main className="hero-bg relative w-full space-y-5 bg-fixed pr-5">
+      <main className="relative w-full space-y-5 bg-fixed pr-5">
         <PageHeader />
         {children}
       </main>
