+++ components/sidebar/SideNav.js
@@ -1,0 +1,68 @@
+import { List } from "@material-tailwind/react";
+import { SwatchIcon, CurrencyEuroIcon } from "@heroicons/react/24/outline";
+
+import SideNavList from "./SideNavList";
+
+const SideNav = ({ open, handleOpen }) => {
+  const walletSubItems = [
+    {
+      label: "Transfer",
+      href: "/wallet/transfer",
+      icon: (
+        <path d="M16.8281 2.6291C16.6781 2.69004 16.4672 2.89161 16.3687 3.06973C16.2984 3.1916 16.2891 3.41192 16.275 4.60254L16.2609 5.99942L13.1297 6.00879L10.0031 6.02286L9.83438 6.13536C9.17813 6.58536 9.36094 7.52286 10.1391 7.68692C10.3172 7.71973 11.4375 7.73379 13.8797 7.72442C17.6859 7.71036 17.4937 7.72442 17.7516 7.44317C17.9625 7.21817 18 7.03536 18 6.24786V5.50723L19.9547 7.46661L21.9141 9.42129L19.9547 11.376L18 13.3354V12.5947C18 11.6525 17.9297 11.451 17.5219 11.2447C17.3531 11.1604 17.1656 11.1557 12.5437 11.1557H7.73906L7.725 9.75879C7.71094 8.41817 7.70625 8.35723 7.60781 8.16973C7.41094 7.80879 6.92813 7.63536 6.52969 7.78067C6.33281 7.85567 0.201563 13.9916 0.0890633 14.226C-0.0234367 14.465 -0.0234367 14.69 0.0890633 14.9291C0.150001 15.0557 1.16719 16.1057 3.28594 18.215C6.63281 21.5525 6.525 21.4588 6.97031 21.4072C7.20938 21.3838 7.47188 21.1963 7.61719 20.9479C7.70625 20.7979 7.71094 20.69 7.725 19.3963L7.74375 17.9994L10.8703 17.99L13.9969 17.976L14.1656 17.8635C14.8219 17.4135 14.6391 16.476 13.8609 16.3119C13.6828 16.2791 12.5625 16.265 10.1203 16.2744C6.31875 16.2885 6.50156 16.2744 6.25313 16.5557C6.04219 16.7807 6.02344 16.8744 6 17.6854L5.97656 18.4682L4.01719 16.5135L2.0625 14.5541L4.03125 12.5854L6 10.6166V11.3807C6 12.351 6.06563 12.5432 6.47813 12.7541C6.64688 12.8385 6.83438 12.8432 11.4563 12.8432H16.2609L16.275 14.2354C16.2891 15.5807 16.2937 15.6416 16.3922 15.8291C16.5891 16.19 17.0719 16.3635 17.4703 16.2182C17.6672 16.1432 23.7984 10.0072 23.9109 9.77286C24.0187 9.54317 24.0234 9.33223 23.9297 9.10254C23.8266 8.86348 17.775 2.79317 17.5078 2.6666C17.3062 2.57285 17.0156 2.55411 16.8281 2.6291Z" />
+      ),
+      fill: true,
+    },
+    {
+      label: "Deposit",
+      href: "/wallet/deposit",
+      icon: (
+        <path d="M4.26562 2.09011C3.43594 2.30573 2.69531 2.91042 2.32031 3.67917C1.99219 4.34011 2.01094 3.77292 2.025 12.1214L2.03906 19.6635L2.16094 19.9776C2.50781 20.8682 3.13125 21.4917 4.02187 21.8385L4.33594 21.9604H7.45312C10.5422 21.9604 10.575 21.9604 10.8984 21.8573C11.7797 21.576 12.4828 20.8964 12.8391 19.9776L12.9609 19.6635L12.975 17.0432L12.9891 14.4229L13.6781 15.1073C14.1328 15.5573 14.4422 15.8245 14.5781 15.8854C15.4359 16.2745 16.2891 15.3932 15.8672 14.5495C15.7172 14.2448 12.675 11.1979 12.4125 11.0854C12.1687 10.987 11.7891 10.9964 11.5594 11.1089C11.2687 11.2495 8.23594 14.301 8.11406 14.5776C7.73437 15.4214 8.58281 16.2698 9.42187 15.8854C9.55781 15.8245 9.86719 15.5573 10.3219 15.1026L11.0156 14.4182V16.7292C11.0156 18.3417 10.9969 19.101 10.9594 19.2557C10.8844 19.5557 10.5562 19.8839 10.2562 19.9589C9.94687 20.0385 5.05312 20.0385 4.74375 19.9589C4.44375 19.8839 4.11562 19.5557 4.04062 19.2557C3.975 19.0073 3.95625 7.82761 4.02187 7.82761C4.04062 7.82761 4.1625 7.86511 4.28906 7.91198C4.50937 7.98698 5.05781 7.99167 12.2719 8.00573L20.0156 8.0198L20.0063 14.0057L19.9922 19.9917L17.8359 20.0151C15.4781 20.0385 15.5531 20.0292 15.2719 20.3292C14.8125 20.812 14.9578 21.5807 15.5766 21.8948C15.7453 21.9839 15.8625 21.9839 18.5531 21.9745L21.3609 21.9604L21.5437 21.8339C21.6422 21.7635 21.7781 21.6276 21.8437 21.5292L21.9609 21.3557L21.975 14.0526C21.9844 6.9323 21.9844 6.7448 21.8953 6.57136C21.7781 6.34636 21.5859 6.16355 21.3656 6.07448C21.2109 6.00886 20.2969 5.99948 12.9937 5.99948H4.8L4.55156 5.87761C3.9375 5.56823 3.81562 4.74323 4.31719 4.29323C4.42031 4.19948 4.56094 4.09636 4.63125 4.06355C4.72969 4.02136 6.72656 4.00261 13.0594 3.98386L21.3562 3.96042L21.5297 3.84324C21.6281 3.77761 21.7641 3.64167 21.8344 3.54323C21.9469 3.37917 21.9609 3.31355 21.9609 2.99949C21.9609 2.68542 21.9469 2.6198 21.8344 2.45573C21.7641 2.3573 21.6281 2.22136 21.5297 2.15574L21.3562 2.03855L12.9422 2.02917C5.72344 2.02448 4.48594 2.02917 4.26562 2.09011Z" />
+      ),
+      fill: true,
+    },
+    {
+      label: "Swap",
+      href: "/wallet/swap",
+      icon: (
+        <svg
+          xmlns="http://www.w3.org/2000/svg"
+          width="24"
+          height="24"
+          viewBox="0 0 24 24"
+        >
+          <path
+            d="M17 5.12501C19.1213 6.67054 20.5 9.17407 20.5 11.9996C20.5 16.694 16.6944 20.4996 12 20.4996H11.5M7 18.8742C4.87867 17.3287 3.5 14.8251 3.5 11.9996C3.5 7.30519 7.30558 3.49961 12 3.49961H12.5M13 22.3996L11 20.3996L13 18.3996M11 5.59961L13 3.59961L11 1.59961"
+            strokeWidth="1.5"
+            strokeLinecap="round"
+            strokeLinejoin="round"
+          />
+        </svg>
+      ),
+      fill: false,
+    },
+  ];
+
+  const assetsSubItems = [
+    { label: "NFTs", href: "/assets/NFTs", icon: <SwatchIcon />, fill: false },
+    {
+      label: "Tokens",
+      href: "/assets/tokens",
+      icon: <CurrencyEuroIcon />,
+      fill: false,
+    },
+  ];
+
+  return (
+    <List>
+      <SideNavList
+        open={open}
+        handleOpen={handleOpen}
+        walletSubItems={walletSubItems}
+        assetsSubItems={assetsSubItems}
+      />
+    </List>
+  );
+};
+
+export default SideNav;
