+++ components/ui/buttons/DefaultButton.js
@@ -6,7 +6,7 @@ const DefaultButton = ({ children, style, icon, ...props }) => {
   return (
     <Button
       className={`font-grotesque ${style} ${
-        icon && " flex items-center justify-center gap-3"
+        icon && " flex items-center justify-center gap-2"
       }`}
       {...props}
     >
